move pipelines into submodule
